feat(pagination)!: add scoped opaque list cursors
Replace exposed numeric continuation boundaries across Guild, User, and Message list RPCs with HMAC-SHA256-signed opaque tokens. Bind each token to its list kind, resource, user, role, and optional filter scope so valid cursors cannot be replayed against a different query.

Adopt stable keyset ordering with limit-plus-one lookahead so next_cursor is present only when another page exists. Add matching PostgreSQL indexes and explicit BIGINT cursor casts for Snowflake IDs and millisecond timestamps.

Wire cursor signing configuration through the User, Guild, Message, API, Dispatcher, and compose paths. Regenerate protobuf bindings and update English and Chinese documentation plus unit, composition, store, and integration coverage.

BREAKING CHANGE: list pagination request and response fields now use opaque cursor strings instead of numeric boundary IDs.

+++ services/message/v1/internal/store/queries.go
@@ -219,7 +219,7 @@ const listDmChannelsQuery = `
 	SELECT id, user_lo, user_hi, created_at
 	FROM dm_channels
 	WHERE (user_lo = $1 OR user_hi = $1)
-	  AND ($2 = 0 OR id < $2)
+	  AND ($2::BIGINT = 0 OR id < $2::BIGINT)
 	ORDER BY id DESC
 	LIMIT $3
 `
